+++ pkg/providers/appsec/testdata/TestResAdvancedSettingsPIILearning/match_by_id.tf
@@ -1,0 +1,10 @@
+provider "akamai" {
+  edgerc        = "../../test/edgerc"
+  cache_enabled = false
+}
+
+resource "akamai_appsec_advanced_settings_pii_learning" "test" {
+  config_id           = 43253
+  enable_pii_learning = true
+}
+